requirements.txt
auth0-python
python-dotenv